Releases/2.7.13 (#257)
Bug fixes:
  - config: only run arg_parse function if executing from CLI or test suite (i.e. scripts shouldn't inadvertently trigger arg_parse when reusing the build_config function)
  - config: make ARTIST_FIRST a global config option so it can be set via the CLI for both sync and utils operations
  - copying playlists: when multiple playlists exist with the same name, it was possible for one or more of the playlists to be inadvertently removed from the output collection
  - playlist_builder: when ordinary operands and track sets were considered separately, the information about the order in which the different operand types appeared got lost resulting in incorrect evaluation
  - playlist_builder: lower precision date inequalities weren't evaluated properly (e.g. {date:>2022} would match a track added any time after 2022-01-01 because `2022` converted into a datetime object becomes 2022-01-01)
  - playlist_builder: globbing strings with * characters would match substrings as if implicit ^ and $ characters weren't being used -- use `match` instead of `search` and add implicit `$` character
  - playlist_buidler: MinimalDeepTechFilter should only be applied to playlists underneath a "Techno" or "House" playlist
  - sync: upload_log logic fix for which files are removed from the logs directory
  - sync: handle errors decoding bytes from stdout when syncing (replace bytes that fail decoding with empty strings)
  - Fixes for Python 3.8 - 3.12 and Windows compatibility
CI:
  - Pytest and pylint with the latest Python versions for 3.8, 3.9, 3.10, 3.11, and 3.12
  - Pytest and pylint on Windows in addition to Ubuntu
Docs:
  - Add tagging guide
Enhancements:
  - playlist_builder: add filtering and aggregation steps to combiner playlists in addition to tags playlists
  - playlist_builder: rename "Other" playlist / folder to "Unused Tags"
  - playlisb_guilder: remove pre-existing playlist_builder playlists
  - playlist_builder: gracefully handle missing playlists when using playlist selectors 
  - playlist_builder: gracefully handle exceptions during parsing of combiner expressions
  - utils: parallelize audio export in process_recording
Features:
  - playlist_builder: configurable PlaylistName (allows playlist names to be different from their combiner expressions or tags)
  - playlist_builder: TransitionTrackFilter and ComplexTrackFilter (playlists from tracks whose comments contain special syntax for BPM or genre transitions, playlists containing tracks with many "other" tags)
  - process_recording: trim shift heuristic to automatically remove leading silence of recording
Refactor:
  - collections: split abstractions and implementations for Collection, Playlist, and Track into their own modules
Scripts:
  - cluster_tracks.py: generate playlists by clustering tracks on tag data
  - get_tags_from_spotify.py: add interactive mode and --mode flag to switch between interactive and bulk processing
  - update_collection.py: sync local Collection data to a remote master Collection

+++ src/djtools/sync/helpers.py
@@ -103,7 +103,7 @@ def rewrite_track_paths(config: BaseConfig, other_user_collection: Path):
             music_path / loc.split(str(music_path) + "/", maxsplit=-1)[-1]
         )
         track.set_location(config.USB_PATH / common_path)
-    collection.serialize(output_path=other_user_collection)
+    collection.serialize(path=other_user_collection)
 
 
 def run_sync(_cmd: str) -> str:
@@ -115,6 +115,7 @@ def run_sync(_cmd: str) -> str:
 
     Raises:
         CalledProcessError: raised if "aws s3 sync" command fails.
+        RuntimeError: raised if any other exception occurs while syncing.
 
     Returns:
         Formatted list of uploaded tracks; tracks are grouped by directory.
@@ -125,7 +126,10 @@ def run_sync(_cmd: str) -> str:
     try:
         with Popen(_cmd, stdout=PIPE) as proc:
             while True:
-                char = proc.stdout.read(1).decode()
+                try:
+                    char = proc.stdout.read(1).decode()
+                except UnicodeDecodeError:
+                    char = ""
                 if char == "" and proc.poll() is not None:
                     break
                 if char not in termination_chars:
@@ -143,7 +147,7 @@ def run_sync(_cmd: str) -> str:
     except Exception as exc:
         msg = f"Failure while syncing: {exc}"
         logger.critical(msg)
-        raise Exception(msg) from exc
+        raise RuntimeError(msg) from exc
 
     new_music = ""
     for group_id, group in groupby(
@@ -187,9 +191,11 @@ def upload_log(config: BaseConfig, log_file: Path):
     now = datetime.now()
     one_day = timedelta(days=1)
     for _file in log_file.parent.rglob("*"):
-        if _file.name == "__init__.py" or not _file.is_file():
-            continue
-        if _file.lstat().st_mtime < (now - one_day).timestamp():
+        if (
+            _file.name != "__init__.py"
+            and _file.is_file()
+            and _file.lstat().st_mtime < (now - one_day).timestamp()
+        ):
             _file.unlink()
